requirements.txt
crewai>=0.1.0
langchain>=0.0.300
langchain-openai>=0.0.1
supabase>=1.0.0
python-dotenv>=1.0.0